chore: refine config code, adjusted the module and dependencies, distinguish between the underlying sdk and the controller sdk
