+++ internal/cli/cmd/cluster/config_observer.go
@@ -41,6 +41,7 @@ import (
 	"github.com/apecloud/kubeblocks/internal/cli/util"
 	"github.com/apecloud/kubeblocks/internal/cli/util/flags"
 	cfgcore "github.com/apecloud/kubeblocks/internal/configuration"
+	"github.com/apecloud/kubeblocks/internal/configuration/validate"
 	"github.com/apecloud/kubeblocks/internal/constant"
 )
 
@@ -176,7 +177,7 @@ func (r *configObserverOptions) printExplainConfigure(configSpecs configSpecsTyp
 			fmt.Printf("\n%s\n", fmt.Sprintf(notConfigSchemaPrompt, printer.BoldYellow(tplName)))
 			return nil
 		}
-		apiSchema, err := cfgcore.GenerateOpenAPISchema(schema.CUE, "")
+		apiSchema, err := validate.GenerateOpenAPISchema(schema.CUE, "")
 		if err != nil {
 			return cfgcore.WrapError(err, "failed to generate open api schema")
 		}
