+++ internal/configuration/config_test.go
@@ -301,66 +301,3 @@ func sortParams(param []VisualizedParam) {
 		})
 	}
 }
-
-func TestIsQuotesString(t *testing.T) {
-	tests := []struct {
-		name string
-		arg  string
-		want bool
-	}{{
-		name: "quotes_test",
-		arg:  ``,
-		want: false,
-	}, {
-		name: "quotes_test",
-		arg:  `''`,
-		want: true,
-	}, {
-		name: "quotes_test",
-		arg:  `""`,
-		want: true,
-	}, {
-		name: "quotes_test",
-		arg:  `'`,
-		want: false,
-	}, {
-		name: "quotes_test",
-		arg:  `"`,
-		want: false,
-	}, {
-		name: "quotes_test",
-		arg:  `for test`,
-		want: false,
-	}, {
-		name: "quotes_test",
-		arg:  `'test'`,
-		want: true,
-	}, {
-		name: "quotes_test",
-		arg:  `'test`,
-		want: false,
-	}, {
-		name: "quotes_test",
-		arg:  `test'`,
-		want: false,
-	}, {
-		name: "quotes_test",
-		arg:  `"test"`,
-		want: true,
-	}, {
-		name: "quotes_test",
-		arg:  `test"`,
-		want: false,
-	}, {
-		name: "quotes_test",
-		arg:  `"test`,
-		want: false,
-	}}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := isQuotesString(tt.arg); got != tt.want {
-				t.Errorf("isQuotesString() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
