+++ internal/configuration/config_util_test.go
@@ -20,7 +20,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 package configuration
 
 import (
-	"encoding/json"
 	"reflect"
 	"strings"
 	"testing"
@@ -28,14 +27,10 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/StudioSol/set"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/apecloud/kubeblocks/apis/apps/v1alpha1"
-	cfgutil "github.com/apecloud/kubeblocks/internal/configuration/util"
-	testapps "github.com/apecloud/kubeblocks/internal/testutil/apps"
 	testutil "github.com/apecloud/kubeblocks/internal/testutil/k8s"
-	"github.com/apecloud/kubeblocks/test/testdata"
 )
 
 var _ = Describe("config_util", func() {
@@ -50,105 +45,6 @@ var _ = Describe("config_util", func() {
 	AfterEach(func() {
 		// Add any teardown steps that needs to be executed after each test
 		k8sMockClient.Finish()
-	})
-
-	Context("MergeAndValidateConfigs", func() {
-		It("Should succeed with no error", func() {
-			type args struct {
-				configConstraint v1alpha1.ConfigConstraintSpec
-				baseCfg          map[string]string
-				updatedParams    []ParamPairs
-				cmKeys           []string
-			}
-
-			configConstraintObj := testapps.NewCustomizedObj("resources/mysql-config-constraint.yaml",
-				&v1alpha1.ConfigConstraint{}, func(cc *v1alpha1.ConfigConstraint) {
-					if ccContext, err := testdata.GetTestDataFileContent("cue_testdata/pg14.cue"); err == nil {
-						cc.Spec.ConfigurationSchema = &v1alpha1.CustomParametersValidation{
-							CUE: string(ccContext),
-						}
-					}
-					cc.Spec.FormatterConfig = &v1alpha1.FormatterConfig{
-						Format: v1alpha1.Properties,
-					}
-				})
-
-			cfgContext, err := testdata.GetTestDataFileContent("cue_testdata/pg14.conf")
-			Expect(err).Should(Succeed())
-
-			tests := []struct {
-				name    string
-				args    args
-				want    map[string]string
-				wantErr bool
-			}{{
-				name: "pg1_merge",
-				args: args{
-					configConstraint: configConstraintObj.Spec,
-					baseCfg: map[string]string{
-						"key":  string(cfgContext),
-						"key2": "not support context",
-					},
-					updatedParams: []ParamPairs{
-						{
-							Key: "key",
-							UpdatedParams: map[string]interface{}{
-								"max_connections": "200",
-								"shared_buffers":  "512M",
-							},
-						},
-					},
-					cmKeys: []string{"key", "key3"},
-				},
-				want: map[string]string{
-					"max_connections": "200",
-					"shared_buffers":  "512M",
-				},
-			}, {
-				name: "not_support_key_updated",
-				args: args{
-					configConstraint: configConstraintObj.Spec,
-					baseCfg: map[string]string{
-						"key":  string(cfgContext),
-						"key2": "not_support_context",
-					},
-					updatedParams: []ParamPairs{
-						{
-							Key: "key",
-							UpdatedParams: map[string]interface{}{
-								"max_connections": "200",
-								"shared_buffers":  "512M",
-							},
-						},
-					},
-					cmKeys: []string{"key1", "key2"},
-				},
-				wantErr: true,
-			}}
-			for _, tt := range tests {
-				got, err := MergeAndValidateConfigs(tt.args.configConstraint, tt.args.baseCfg, tt.args.cmKeys, tt.args.updatedParams)
-				Expect(err != nil).Should(BeEquivalentTo(tt.wantErr))
-				if tt.wantErr {
-					continue
-				}
-
-				option := CfgOption{
-					Type:    CfgTplType,
-					CfgType: tt.args.configConstraint.FormatterConfig.Format,
-				}
-
-				patch, err := CreateMergePatch(&ConfigResource{
-					ConfigData: tt.args.baseCfg,
-				}, &ConfigResource{
-					ConfigData: got,
-				}, option)
-				Expect(err).Should(Succeed())
-
-				var patchJSON map[string]string
-				Expect(json.Unmarshal(patch.UpdateConfig["key"], &patchJSON)).Should(Succeed())
-				Expect(patchJSON).Should(BeEquivalentTo(tt.want))
-			}
-		})
 	})
 
 	Context("common funcs test", func() {
@@ -235,50 +131,6 @@ var _ = Describe("config_util", func() {
 	})
 
 })
-
-func TestFromUpdatedConfig(t *testing.T) {
-	type args struct {
-		base map[string]string
-		sets *set.LinkedHashSetString
-	}
-	tests := []struct {
-		name string
-		args args
-		want map[string]string
-	}{{
-		name: "normal_test",
-		args: args{
-			base: map[string]string{
-				"key1": "config context1",
-				"key2": "config context2",
-				"key3": "config context2",
-			},
-			sets: set.NewLinkedHashSetString("key1", "key3"),
-		},
-		want: map[string]string{
-			"key1": "config context1",
-			"key3": "config context2",
-		},
-	}, {
-		name: "none_updated_test",
-		args: args{
-			base: map[string]string{
-				"key1": "config context1",
-				"key2": "config context2",
-				"key3": "config context2",
-			},
-			sets: cfgutil.NewSet(),
-		},
-		want: map[string]string{},
-	}}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := fromUpdatedConfig(tt.args.base, tt.args.sets); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("fromUpdatedConfig() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
 
 func TestMergeUpdatedConfig(t *testing.T) {
 	type args struct {
