+++ internal/configuration/validate/utils.go
@@ -1,0 +1,36 @@
+/*
+Copyright (C) 2022-2023 ApeCloud Co., Ltd
+
+This file is part of KubeBlocks project
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Affero General Public License for more details.
+
+You should have received a copy of the GNU Affero General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+package validate
+
+// isQuotesString checks whether a string is quoted.
+func isQuotesString(str string) bool {
+	const (
+		singleQuotes = '\''
+		doubleQuotes = '"'
+	)
+
+	if len(str) < 2 {
+		return false
+	}
+
+	firstChar := str[0]
+	lastChar := str[len(str)-1]
+	return (firstChar == singleQuotes && lastChar == singleQuotes) || (firstChar == doubleQuotes && lastChar == doubleQuotes)
+}
